news-automation/requirements.txt
feedparser==6.0.11
requests==2.31.0
PyYAML==6.0.1
